+++ src/common/lib/client/realtimechannel.ts
@@ -183,7 +183,20 @@ class RealtimeChannel extends Channel {
   }
 
   _shouldReattachToSetOptions(options?: API.Types.ChannelOptions) {
-    return (this.state === 'attached' || this.state === 'attaching') && (options?.params || options?.modes);
+    if (!(this.state === 'attached' || this.state === 'attaching')) {
+      return false;
+    }
+    if (options?.params) {
+      if (!this.params || !Utils.shallowEquals(this.params, options.params)) {
+        return true;
+      }
+    }
+    if (options?.modes) {
+      if (!this.modes || !Utils.arrEquals(options.modes, this.modes)) {
+        return true;
+      }
+    }
+    return false;
   }
 
   publish(...args: any[]): void | Promise<void> {