+++ src/common/lib/util/utils.ts
@@ -307,10 +307,10 @@ export const arrFilter = (Array.prototype.filter as unknown)
     };
 
 export const arrEvery = (Array.prototype.every as unknown)
-  ? function <T>(arr: Array<T>, fn: (value: T, index?: number, arr?: Array<T>) => boolean) {
+  ? function <T>(arr: Array<T>, fn: (value: T, index: number, arr: Array<T>) => boolean) {
       return arr.every(fn);
     }
-  : function <T>(arr: Array<T>, fn: (value: T, index?: number, arr?: Array<T>) => boolean) {
+  : function <T>(arr: Array<T>, fn: (value: T, index: number, arr: Array<T>) => boolean) {
       const len = arr.length;
       for (let i = 0; i < len; i++) {
         if (!fn(arr[i], i, arr)) {
@@ -593,4 +593,13 @@ export function toBase64(str: string) {
   const bufferUtils = Platform.BufferUtils;
   const textBuffer = bufferUtils.utf8Encode(str);
   return bufferUtils.base64Encode(textBuffer);
+}
+
+export function arrEquals(a: any[], b: any[]) {
+  return (
+    a.length === b.length &&
+    arrEvery(a, function (val, i) {
+      return val === b[i];
+    })
+  );
 }